fix(Skeleton): can not register skeleton image component (#11470)
* fix(Skeleton): can not register skeleton image component by default #11469

* chore(skeleton-avatar): migrate skeleton avatar to new folder

* chore: remove duplicate types

* chore(skeleton-image): migrate skeleton image to new folder

* chore(skeleton-paragraph): migrate skeleton paragraph to new folder

* chore(skeleton-title): migrate skeleton title to new folder
