+++ packages/vant/src/skeleton-avatar/README.md
@@ -1,0 +1,3 @@
+# SkeletonAvatar
+
+Please refer to [Skeleton docs](/skeleton#skeletonavatar-props)
