+++ packages/vant/src/skeleton-avatar/SkeletonAvatar.tsx
@@ -8,8 +8,7 @@ import {
   createNamespace,
 } from '../utils';
 
-// Types
-import type { SkeletonAvatarShape } from './types';
+export type SkeletonAvatarShape = 'square' | 'round';
 
 const [name, bem] = createNamespace('skeleton-avatar');
