+++ packages/vant/src/skeleton-avatar/index.less
@@ -1,0 +1,22 @@
+:root {
+  --van-skeleton-avatar-size: 32px;
+  --van-skeleton-avatar-background: var(--van-active-color);
+}
+
+.van-skeleton {
+  &-avatar {
+    flex-shrink: 0;
+    width: var(--van-skeleton-avatar-size);
+    height: var(--van-skeleton-avatar-size);
+    margin-right: var(--van-padding-md);
+    background: var(--van-skeleton-avatar-background);
+
+    &--round {
+      border-radius: var(--van-radius-max);
+    }
+  }
+
+  &-avatar + &__content {
+    padding-top: var(--van-padding-xs);
+  }
+}
