+++ packages/vant/src/skeleton-image/README.md
@@ -1,0 +1,3 @@
+# SkeletonImage
+
+Please refer to [Skeleton docs](/skeleton#skeletonimage-props)
