+++ packages/vant/src/skeleton-image/SkeletonImage.tsx
@@ -7,12 +7,11 @@ import {
   createNamespace,
 } from '../utils';
 
-// Types
-import type { SkeletonImageShape } from './types';
-
 import { Icon } from '../icon';
 
 const [name, bem] = createNamespace('skeleton-image');
+
+export type SkeletonImageShape = 'square' | 'round';
 
 export const skeletonImageProps = {
   imageSize: numericProp,
