+++ packages/vant/src/skeleton-paragraph/README.md
@@ -1,0 +1,3 @@
+# SkeletonParagraph
+
+Please refer to [Skeleton docs](/skeleton#skeletonparagraph-props)
