+++ packages/vant/src/skeleton-paragraph/index.ts
@@ -1,0 +1,14 @@
+import _SkeletonParagraph from './SkeletonParagraph';
+import { withInstall } from '../utils';
+
+export const SkeletonParagraph = withInstall(_SkeletonParagraph);
+export default SkeletonParagraph;
+
+export { skeletonParagraphProps, DEFAULT_ROW_WIDTH } from './SkeletonParagraph';
+export type { SkeletonParagraphProps } from './SkeletonParagraph';
+
+declare module 'vue' {
+  export interface GlobalComponents {
+    VanSkeletonParagraph: typeof SkeletonParagraph;
+  }
+}
