+++ packages/vant/src/skeleton-title/README.md
@@ -1,0 +1,3 @@
+# SkeletonParagraph
+
+Please refer to [Skeleton docs](/skeleton#skeletontitle-props)
