+++ packages/vant/src/skeleton/types.ts
@@ -1,7 +1,3 @@
-export type SkeletonAvatarShape = 'square' | 'round';
-
-export type SkeletonImageShape = 'square' | 'round';
-
 export type SkeletonThemeVars = {
   skeletonParagraphHeight?: string;
   skeletonParagraphBackground?: string;
